flask-app/requirements.txt
Flask==3.0.0
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-api
opentelemetry-distro
